+++ docs/source/news.rst
@@ -8,6 +8,9 @@ Changelog
 - fix: prevent raising :exc:`AttributeError` when ``--reload`` is not passed
   in case of a :exc:`SyntaxError` raised from the WSGI application.
   (:issue:`1805`, :pr:`1806`)
+- The internal module ``gunicorn.workers.async`` was renamed to ``gunicorn.workers.base_async``
+  since ``async`` is now a reserved word in Python 3.7.
+  (:pr:`1527`)
 
 19.8.1 / 2018/04/30
 ===================